# backend/src/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Swagger for a Node.js Api about superheroes",
        "version": "1.0.0",
        "title": "Swagger MyHero",
        "contact": {
            "email": "m.mercedes.perea@gmail.com"
        }
    },
    "host": "backendmyhero.eu-gb.mybluemix.net",
    "basePath": "/api",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/user/checkFollow/{idUsu}/{idUnfollow}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "check whether a user is being followed",
                "description": "check whether a user is being followed",
                "operationId": "checkFollow",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idUsu",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "idUnfollow",
                        "in": "path",
                        "description": "id user follow",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Ids must be sent"
                    }
                }
            }
        },
        "/user/getFollowUsers/{idUsu}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get all the users you follow",
                "description": "Get all the users you follow",
                "operationId": "getFollowUsers",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idUsu",
                        "in": "path",
                        "description": "id User",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "name must be sent"
                    }
                }
            }
        },
        "/user/getFollowersUsers/{idUsu}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get all the users who follow you",
                "description": "Get all the users who follow you",
                "operationId": "getFollowersUsers",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idUsu",
                        "in": "path",
                        "description": "id User",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "name must be sent"
                    }
                }
            }
        },
        "/user/{idUsu}": {
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "update a user",
                "description": "update a user",
                "operationId": "updateUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idUsu",
                        "in": "path",
                        "description": "id User",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "update a user",
                        "required": [
                            "email",
                            "userName",
                            "alias",
                            "surname",
                            "dateOfBirth"
                        ],
                        "properties": {
                            "email": {
                                "type": "string",
                                "desciption": "email"
                            },
                            "userName": {
                                "type": "string",
                                "desciption": "userName"
                            },
                            "alias": {
                                "type": "string",
                                "desciption": "email"
                            },
                            "surname": {
                                "type": "string",
                                "desciption": "surname"
                            },
                            "dateOfBirth": {
                                "type": "string",
                                "format": "date-time",
                                "desciption": "date Of Birth"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "ids must be sent"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/user/deleteUser/{idUsu}": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Delete a user",
                "description": "Delete a user",
                "operationId": "DeleteUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idUsu",
                        "in": "path",
                        "description": "id User",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "comment a hero",
                        "required": [
                            "email",
                            "password"
                        ],
                        "properties": {
                            "email": {
                                "type": "string",
                                "desciption": "email"
                            },
                            "password": {
                                "type": "string",
                                "desciption": "old password"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/user/followUser": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Follow a user",
                "description": "Follow a user",
                "operationId": "followUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "comment a hero",
                        "required": [
                            "idUsu",
                            "idUsuFollow"
                        ],
                        "properties": {
                            "idUsu": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "idUsuFollow": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/user/unFollowUser/{idUsu}/{idUnfollow}": {
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Unfollow a user",
                "description": "Unfollow a user",
                "operationId": "unFollowUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idUsu",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "idUnfollow",
                        "in": "path",
                        "description": "id user Unfollow",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "team not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "properties": {
                "idUsu": {
                    "type": "integer",
                    "format": "int64"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                },
                "alias": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "admin": {
                    "type": "string"
                }
            }
        }
    }
}